New version of the scraper
I wanted to focus more on making each piece of data scraped be more modular. I reworked the structure, so it is easier to take individual data points by having individual methods for each. Could be better but I wanted to pander to pandas a bit more with this one.