Progress and an annoying bug
Added a progress meter for long scrap sessions. The for loop that I had initially was skipping over an index if I wanted to delete multiple, which came up when there was a "map advantage" awarded to a team, causing one of the maps to be TBD, that case has been handled hopefully.